Create CODEOWNERS
I created the CODEOWNERS file to protect the production branch. Initially, the master developer and Alex Ament have global code ownership over production, so any pull requests adding to the official competition production code must be approved by one of those two.

It is expected that as team members learn the code and create branches, code ownership can be delegated to others and this file will be edited to include that code ownership.

Signed-off-by: Ignatius Master Developer <masterdevsihs@gmail.com>